requirements.txt
beartype==0.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
flet==0.2.4
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
PyPDF2==3.0.1
repath==0.9.0
requests==2.28.1
six==1.16.0
urllib3==1.26.13
watchdog==2.2.1
websocket-client==1.4.2
